requirements.txt
absl-py==2.4.0
astunparse==1.6.3
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.62.0
gast==0.7.0
google-auth==2.49.0
google-auth-oauthlib==1.3.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
idna==3.11
joblib==1.5.3
keras==3.12.1
kiwisolver==1.5.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.4
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
pillow==12.1.1
protobuf==3.20.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sounddevice==0.5.5
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==1.14.2
